requirements.txt
h2o_wave==0.25.3
pandas==2.0.1
black==22.3.0
flake8==6.0.0
pyyaml==6.0